pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wtf"
version = "0.1.0"
description = "Convert natural language to shell commands using AI"
authors = [
    {name = "Meji Abidoye", email ="pfbyjy@example.com"},
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["cli", "shell", "ai", "natural-language"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: System :: Shells",
]
dependencies = [
    "click",
    "pyyaml",
    "openai",
    "anthropic",
    "rich",
    "pyperclip",
]

[project.urls]
Homepage = "https://github.com/pfbyjy/wtf"
Repository = "https://github.com/pfbyjy/wtf.git"

[project.scripts]
wtf = "wtf:main"
wtf-config = "wtf.config_cmd:cli"
